Added configurable headings level
Currently, there was no way for changing level of the heading, from
which html would be generated. In this patch, there is added an ability
to define it in jekyll _config.yml file for site.

It's simple as adding following entries into _config.yml:

rst:
  initial_header_level: 2

If not defined, level 2 will be used as default.